train/requirements.txt
setuptools
appdirs
certifi
cityscapesScripts
coloredlogs
cycler
filelock
humanfriendly
idna
kiwisolver
numpy
packaging
Pillow
pycocotools
pyparsing
pyquaternion
python-dateutil
PyYAML
scipy
torch
brevitas==0.9.1
torchvision
triton
typing
typing_extensions
urllib3
yacs
